+++ src/main.rs
@@ -98,9 +98,11 @@ macro_rules! command_list {
     to            Convert a CSV file to a variety of data formats
     scrape        Scrape HTML into CSV data
     reverse       Reverse rows of CSV data
-    transpose (t) Transpose CSV file
-    pivot         Stack multiple columns into fewer columns
-    unpivot       Split distinct values of a column into their own columns
+
+## Transpose & pivot
+    transpose (t)            Transpose CSV file
+    pivot     (pivot-wider)  Stack multiple columns into fewer columns
+    unpivot   (pivot-longer) Split distinct values of a column into their own columns
 
 ## Split a CSV file into multiple
     split       Split CSV data into chunks
@@ -311,6 +313,10 @@ enum Command {
     Parallel,
     Partition,
     Pivot,
+    #[serde(rename = "pivot-longer")]
+    PivotLonger,
+    #[serde(rename = "pivot-wider")]
+    PivotWider,
     Plot,
     Progress,
     Range,
@@ -395,7 +401,7 @@ impl Command {
             Command::Merge => cmd::merge::run(argv),
             Command::Parallel | Command::P => cmd::parallel::run(argv),
             Command::Partition => cmd::partition::run(argv),
-            Command::Pivot => cmd::pivot::run(argv),
+            Command::Pivot | Command::PivotWider => cmd::pivot::run(argv),
             Command::Plot => cmd::plot::run(argv),
             Command::Progress => cmd::progress::run(argv),
             Command::Range => cmd::range::run(argv),
@@ -416,7 +422,7 @@ impl Command {
             Command::Top => cmd::top::run(argv),
             Command::Transform => cmd::transform::run(argv),
             Command::Transpose | Command::T => cmd::transpose::run(argv),
-            Command::Unpivot => cmd::unpivot::run(argv),
+            Command::Unpivot | Command::PivotLonger => cmd::unpivot::run(argv),
             Command::View | Command::V => cmd::view::run(argv),
             Command::Vocab => cmd::vocab::run(argv),
             Command::Window => cmd::window::run(argv),